Replace health check with `wait-for-services` script (#6)
The job of `wait-for-services` script is to wait until all startup scripts
in `/etc/my_init.d` where executed. It does not wait, however, until services
are properly started (there is no way to make it general). Therefore, if one
wants to achieve that also, one needs to add in the startup script a state
checker.

This PR also adds a startup script that runs at the very last step at startup.
Its job is to create file /root/INIT_COMPLETED signalling that all init
scripts are completed.