move postgresql socket file to /tmp
The current setup placed the postgresql socket file in the home
directory of the aiida user.
In the AiiDAlab, this directory is replaced by a bind mount from the
host.

We have recently encountered a setup (host running MacOS 11.5.2, Docker
Desktop 3.6.0) where postgresql was not able to create the socket file
and failed with [1]

   LOG:  could not bind Unix address "/home/aiida/.postgresql/.s.PGSQL.5432": Invalid argument

Moving the socket file location to /tmp fixes the issue.
Since the socket file is anyhow tied to a specific execution of
postgresql, there should be no downsides to this move.

[1] https://github.com/aiidalab/aiidalab-docker-stack/issues/194#issuecomment-918325330